Fix Firebase build error by renaming firebaseAdmin.ts
The build was failing due to a strange caching issue with the filename `firebaseAdmin.ts`. Even after correcting the file content, the build process was still using an old, cached version of the file.

To work around this issue, this commit renames `firebaseAdmin.ts` to `firebaseConfig.ts` and updates all the imports to point to the new file. This resolves the build error.
