+++ app/api/admin/migrate-urls/route.ts
@@ -1,7 +1,7 @@
 // app/api/admin/migrate-urls/route.ts - Emergency fix for URL document IDs
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { initializeFirebaseAdmin } from '@/lib/firebaseAdmin';
+import { initializeFirebaseAdmin } from '@/lib/firebaseConfig';
 
 export const dynamic = 'force-dynamic';
