--- lib/firebaseAdmin.js
@@ -1,53 +1,0 @@
-"use strict";
-var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    var desc = Object.getOwnPropertyDescriptor(m, k);
-    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
-      desc = { enumerable: true, get: function() { return m[k]; } };
-    }
-    Object.defineProperty(o, k2, desc);
-}) : (function(o, m, k, k2) {
-    if (k2 === undefined) k2 = k;
-    o[k2] = m[k];
-}));
-var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
-    Object.defineProperty(o, "default", { enumerable: true, value: v });
-}) : function(o, v) {
-    o["default"] = v;
-});
-var __importStar = (this && this.__importStar) || (function () {
-    var ownKeys = function(o) {
-        ownKeys = Object.getOwnPropertyNames || function (o) {
-            var ar = [];
-            for (var k in o) if (Object.prototype.hasOwnProperty.call(o, k)) ar[ar.length] = k;
-            return ar;
-        };
-        return ownKeys(o);
-    };
-    return function (mod) {
-        if (mod && mod.__esModule) return mod;
-        var result = {};
-        if (mod != null) for (var k = ownKeys(mod), i = 0; i < k.length; i++) if (k[i] !== "default") __createBinding(result, mod, k[i]);
-        __setModuleDefault(result, mod);
-        return result;
-    };
-})();
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.initializeFirebaseAdmin = initializeFirebaseAdmin;
-const admin = __importStar(require("firebase-admin"));
-const node_buffer_1 = require("node:buffer");
-function initializeFirebaseAdmin() {
-    if (admin.apps.length) {
-        return admin.firestore();
-    }
-    const serviceAccountBase64 = process.env.FIREBASE_ADMIN_SDK_JSON_BASE64;
-    if (!serviceAccountBase64) {
-        throw new Error('FIREBASE_ADMIN_SDK_JSON_BASE64 env variable not set.');
-    }
-    const serviceAccountJson = node_buffer_1.Buffer.from(serviceAccountBase64, 'base64').toString('ascii');
-    const serviceAccount = JSON.parse(serviceAccountJson);
-    admin.initializeApp({
-        credential: admin.credential.cert(serviceAccount),
-    });
-    return admin.firestore();
-}
